Add fuzz harnesses for the untrusted-byte parsers
Repo-rigor pass 3. Three cargo-fuzz targets exercise the surfaces that
parse bytes we don't control:

- sse_scan: upstream SSE fed whole and re-fragmented at an input-derived
  chunk size, asserting the 1 MiB pathological-line guard holds.
- sanitize_label: asserts the output invariants that are the
  metric-injection defense (non-empty, <=64 chars, safe charset).
- config_roundtrip: operator-edited config.json must parse without
  panicking, and serialize -> parse -> serialize must be a fixpoint.

To let the fuzz crate link the internals, src/main.rs became a 3-line
shim over src/lib.rs (nim_proxy::run()); all crate:: paths survived
unchanged and the full suite (69 unit + 53 e2e) is green. The fuzz
wrappers are #[cfg(fuzzing)] + #[doc(hidden)], so normal builds, the
coverage gate, and the shipped binary carry nothing new.

fuzz.yml smoke-fuzzes each target 60s weekly, on dispatch, and on PRs
touching the fuzzed code; crash reproducers upload as artifacts. Seed
corpora live in fuzz/seeds/ (binary in .gitattributes); fuzz/corpus/ is
the gitignored working corpus. Local validation: 977k/12.3M/6.9M
executions per target in 60s, zero findings.
