+++ fuzz/.gitignore
@@ -1,0 +1,5 @@
+target/
+artifacts/
+coverage/
+Cargo.lock
+corpus/
