+++ fuzz/seeds/config_roundtrip/minimal.json
@@ -1,0 +1,1 @@
+{"version":1}
