+++ fuzz/seeds/config_roundtrip/store.json
@@ -1,0 +1,1 @@
+{"version":1,"upstream":{"base_url":"https://integrate.api.nvidia.com","nim_keys":[{"key":"nvapi-test","owner":"admin","enabled":true,"rpm":40}]},"client_auth":{"mode":"keyed","keys":[{"name":"opencode","secret_sha256":"0000000000000000000000000000000000000000000000000000000000000000","owner":"admin"}]},"limits":{"max_wait_secs":30,"heartbeat_secs":10,"stream_idle_secs":120,"request_timeout_secs":300,"max_inflight":64,"strict_passthrough":false},"users":[{"username":"admin","password_hash":"pbkdf2-sha256$1000$00$00","role":"superuser"}]}
