+++ fuzz/seeds/sanitize_label/hostile
@@ -1,0 +1,1 @@
+{__proto__}é🔥<script>"
