+++ fuzz/seeds/sse_scan/malformed
@@ -1,0 +1,4 @@
+data: {"usage":null}
+data: not-json-at-all
+
+
