+++ fuzz/seeds/sse_scan/truncated-mid-json
@@ -1,0 +1,3 @@
+: heartbeat
+
+data: {"choices":[{"delta":{"content":"partial
